Refaktorerer sertifikatlesing
Tilbyr nå to statiske metoder som sørger for å gjøre logging litt
mer ryddig
